Reorganize query module in anticipation of adding join
